# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/checkin-responses": {
      "get": {
        "operationId": "CheckinResponsesController_findAll",
        "parameters": [
          {
            "name": "checkinId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expertId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckinResponseEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "List check-in responses",
        "tags": [
          "checkin-responses"
        ]
      },
      "post": {
        "operationId": "CheckinResponsesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckinResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for the expert response payload."
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Create a check-in response",
        "tags": [
          "checkin-responses"
        ]
      }
    },
    "/checkin-responses/{id}": {
      "get": {
        "operationId": "CheckinResponsesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Check-in response not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Get a check-in response by id",
        "tags": [
          "checkin-responses"
        ]
      },
      "patch": {
        "operationId": "CheckinResponsesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCheckinResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for the expert response payload."
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Check-in response not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Update a check-in response",
        "tags": [
          "checkin-responses"
        ]
      },
      "delete": {
        "operationId": "CheckinResponsesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Check-in response not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Delete a check-in response",
        "tags": [
          "checkin-responses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wellness Platform API",
    "description": "REST APIs for the evaluation backend. Authorization is enforced using the request header role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role required for RBAC. Supported values: Admin, HR, Employee, Wellness Expert"
      }
    },
    "schemas": {
...
      "CheckinResponseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "response-1777718400000-a1b2c3d4"
          },
          "checkinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "checkinType": {
            "type": "string",
            "example": "mental"
          },
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "message": {
            "type": "string",
            "example": "Please take a short break and let us follow up tomorrow."
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-02T12:45:00.000Z"
          }
        },
        "required": [
          "id",
          "checkinId",
          "checkinType",
          "employeeId",
          "employeeName",
          "expertId",
          "expertName",
          "companyId",
          "companyName",
          "message",
          "createdAt"
        ]
      },
      "CreateCheckinResponseDto": {
        "type": "object",
        "properties": {
          "checkinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "checkinType": {
            "type": "string",
            "example": "mental"
          },
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "message": {
            "type": "string",
            "example": "Please take a short break and let us follow up tomorrow."
          }
        },
        "required": [
          "checkinId",
          "checkinType",
          "employeeId",
          "message"
        ]
      },
      "UpdateCheckinResponseDto": {
        "type": "object",
        "properties": {
          "checkinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "checkinType": {
            "type": "string",
            "example": "mental"
          },
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "message": {
            "type": "string",
            "example": "Please take a short break and let us follow up tomorrow."
          }
        }
      },
...
    }
  }
}