# back-end/src/checkin-responses/dto/update-checkin-response.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateCheckinResponseDto } from './create-checkin-response.dto';

export class UpdateCheckinResponseDto extends PartialType(
  CreateCheckinResponseDto,
) {}